+++ pulsr/fileCache.js
@@ -5,12 +5,13 @@
  * in the cache which is served to other modules upon request.
  * async-cache allows to significantly decrease I/O hits.
  */
-define(['async-cache', 'fs'], function(AsyncCache, fs) {
+define(['async-cache', 'fs', 'conf'], function(AsyncCache, fs, conf) {
     var stats = new AsyncCache({
         // options passed directly to the internal lru cache
-        max: 100,
-        // refresh the cache every 10 minutes
-        maxAge: 1000 * 60 * 10,
+        // indicating how many items to store in a fileCache
+        max: conf.app.cache.fileCache.stats.max,
+        // refresh the cache every maxAge time
+        maxAge: conf.app.cache.fileCache.stats.maxAge,
         // method to load a thing if it's not in the cache.
         // key must be unique in the context of this cache.
         load: function (filePath, cb) {