+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Pulsr Web App Framework Change Log
 
+## Version 0.0.9 December 18, 2012
+
+- A fileCache module which caches files stat information. Uses async-cache module.
+- Use fileCache module to get file stat information.
+- Allow users to configure fileCache settings.
+
 ## Version 0.0.8 December 18, 2012
 
 Use Streams to stream static files instead of reading into memory.